rename request-and-parse, change header functionality, format parsing test
